+++ src/main/java/com/example/orange/repository/ParameterRespository.java
@@ -3,5 +3,5 @@ package com.example.orange.repository;
 import com.example.orange.entities.Parameter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParameterRespository extends JpaRepository<Parameter, Long> {
+public interface ParameterRespository extends JpaRepository<Parameter, Integer> {
 }
